Updated/Non-negative Fibonacci

+++ fibonacci.c
@@ -2,7 +2,15 @@
 
 int main() 
 {
-    int fib[100];
+    double fib[100];
+	/*Changing the type of the values to something
+ * 	other than an int provides more precision, and
+ * 	in turn the negatives are not a concern because
+ * 	they are no longer there.
+ * 	Changing the declared type of the fib value and
+ * 	the print statement to %f and instert a float 
+ * 	rather than an int takes care of the alterations*/
+
     int i;
 
     for (i=0; i<100; i++)
@@ -15,7 +23,7 @@ int main()
         {
         fib[i] = fib[i-1] + fib[i-2];
         }
-      printf("%d\n", fib[i]);
+      printf("%f\n", fib[i]);
     }
       
   return 0;
